Add dataParams to NetworkingRequest
Currently, `Params` is a typealias set to
[String: CustomStringConvertible]. This is fine for most cases,
however, there are situations where we may want to just define some
raw data as our HTTP request body.

An example of such an instance is when we want to pass an array as an
argument instead, which is considered non-standard but possible.

Here, we add `dataParams` as a property of `NetworkingRequest`, which
is optionally populated upstream. If `dataParams` exists, we set that
as our request’s `httpBody` instead of attempting to serialize
[String: CustomStringConvertible].