Support filtering words in NetworkingLogger
First, we make sure that NetworkingLogger isn’t initialized by default
in NetworkingClient and Networking Request.

In `NetworkingClient`, you can pass in `filteredWords` as a property in
the initializer that is then used to initialized `NetworkingLogger` with
your defined list of words to filter.

In `NetworkingRequest`, the logger is also parameterized into the
initializer, which initializes to the default implementation of
`NetworkingLogger.

When `NetworkingClient` creates requests, though, it injects the logger
it holds to `NetworkingRequest`.